requirements.txt
gradio==5.49.1
openai>=1.12.0
pandas>=2.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
psycopg2-binary>=2.9.0
argparse

# API
fastapi
uvicorn
